Read from JSON provided by Etsy API
Line 28, the "<script src=...", is where the magic happens. In the URL you'll see that I set the limit=1. That means it will only return 1 result. You can change that to include as many as you want, but higher limits means more requests from ETSY. Of course, when the site goes live, we will remove that limit.